+++ icons/bigCheck.ts
@@ -1,4 +1,4 @@
-import { Icon } from "../utils/type/Icon";
+import { Icon } from "../libs/type/Icon";
 
 const bigCheck: Icon = {
   d: "M9.79,22.23.75,13a2.63,2.63,0,0,1,0-3.68,2.53,2.53,0,0,1,3.57,0l0,0,7.42,7.56L27.64.76a2.51,2.51,0,0,1,3.56,0l.05,0a2.63,2.63,0,0,1,0,3.68L13.77,22.24a2.49,2.49,0,0,1-2,.75h-.19A2.5,2.5,0,0,1,9.79,22.23Z",
